rent_backend/requirements.txt
Django>=4.2,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
django-extensions>=3.2,<4
django-filter>=23.5,<26
python-dotenv>=1.0,<2
